requirements.txt
python == 3.7.0
Django==2.2
django-crispy-forms==1.9.1
django-timezone-utils==0.13
djangorestframework==3.11.0
pytz==2020.1
sqlparse==0.3.1



